Move the address selection bottom sheet to new file
